chore(map-styles): mv tasks to src file
